SoundActionTest updated

+++ actionsystem/src/test/java/com/vabrant/actionsystem/test/tests/ActionSystemTestListener.java
@@ -352,7 +352,7 @@ public class ActionSystemTestListener extends ApplicationAdapter implements Inpu
 			});
 		}
 		
-		public void setValue(int value) {
+		public void setValue(float value) {
 			textField.setText(String.valueOf(value));
 		}
